Применяет паттерн "контейнер" к экрану предложения.
