+++ src/components/offer-screen/offer-screen.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Redirect, useParams} from 'react-router-dom';
+
 import PropTypes from 'prop-types';
 import {reviewPropType, offerPropType} from '../../prop-types';
 
@@ -12,16 +12,7 @@ import {cardTypes, OfferTypes} from '../../const';
 
 
 const OfferScreen = (props) => {
-  const {offers, reviews} = props;
-  const {id} = useParams();
-  const isAuthorized = true; // Temporary
-
-  const offer = offers.find((item) => item.id === +id);
-  const offersNearby = [offers[1], offers[2], offers[3]];
-
-  if (!offer) {
-    return <Redirect to="/" />;
-  }
+  const {offer, reviews, offersNearby, isAuthorized} = props;
 
   return (
     <div className="page">
@@ -117,8 +108,10 @@ const OfferScreen = (props) => {
 };
 
 OfferScreen.propTypes = {
-  offers: PropTypes.arrayOf(offerPropType),
-  reviews: PropTypes.arrayOf(reviewPropType)
+  offer: offerPropType,
+  offersNearby: PropTypes.arrayOf(offerPropType),
+  reviews: PropTypes.arrayOf(reviewPropType),
+  isAuthorized: PropTypes.bool
 };
 
 export default OfferScreen;
